directory refactor
