+++ src/app/modules/inventory-detail/detail/detail.component.html
