+++ src/app/modules/inventory-detail/detail/grid/building-files-grid.component.ts
