+++ src/app/modules/inventory-detail/detail/grid/documents-grid.component.html
