+++ src/app/modules/inventory-detail/detail/grid/history-grid.component.html
