+++ src/app/modules/inventory-detail/detail/grid/paired-grid.component.ts
