+++ src/app/modules/inventory-detail/detail/grid/scenarios-grid.component.html
