+++ src/app/modules/inventory-detail/detail/header.component.html
