+++ src/app/modules/inventory-detail/detail/index.ts
@@ -1,0 +1,8 @@
+export * from './detail.component'
+export * from './grid/building-files-grid.component'
+export * from './grid/documents-grid.component'
+export * from './grid/history-grid.component'
+export * from './grid/paired-grid.component'
+export * from './grid/scenarios-grid.component'
+export * from './header.component'
+export * from './map.component'
