+++ src/app/modules/inventory-detail/detail/map.component.ts
