+++ src/app/modules/inventory-detail/detail/modal/document-upload.component.ts
