+++ src/app/modules/inventory-detail/detail/modal/edit-state.component.html
