+++ src/app/modules/inventory-list/index.ts
@@ -1,2 +1,2 @@
-export * from './grid'
+export * from './list/grid'
 export * from './list/inventory.component'
