+++ src/app/modules/inventory-list/list/grid/actions.component.html
