+++ src/app/modules/inventory-list/list/grid/config-selector.component.html
