+++ src/app/modules/inventory-list/list/grid/config-selector.component.ts
@@ -1,7 +1,7 @@
 import { Component, EventEmitter, Input, Output } from '@angular/core'
 import { MatSelectModule } from '@angular/material/select'
 import type { Cycle } from '@seed/api/cycle'
-import type { Profile } from '../../inventory/inventory.types'
+import type { Profile } from '../../../inventory/inventory.types'
 
 @Component({
   selector: 'seed-inventory-grid-config-selector',
