+++ src/app/modules/inventory-list/list/grid/filter-sort-chips.component.html
