+++ src/app/modules/inventory-list/list/grid/grid-controls.component.html
