+++ src/app/modules/inventory-list/list/grid/grid.component.html
