+++ src/app/modules/inventory-list/list/grid/index.ts
