+++ src/app/modules/inventory-list/list/modal/delete-modal.component.html
