+++ src/app/modules/inventory-list/list/modal/index.ts
