+++ src/app/modules/inventory-list/list/modal/more-actions-modal.component.html
